Change-requests resolved
